Adjust how end-to-end tests are handled.
- We rename "functional tests" to "end-to-end" tests becuase we think
it is a more descriptive name for what the tests do.

- We make a new test category for the end-to-end tests. Previously these
were lumped into the go tests.

- We set the default to not run the end-to-end tests. This means that the
end-to-end tests will not be run on the build machine. This is what we
want for now because they are flaky. In the future when we make them
less flaky we will change the default again. At any rate it is useful
to be able to run them or not with a flag.

Change-Id: Ie661352a0bf7de744808762bbf1d33af97a04317