PKE updates to improve bandwidth, fix flaky tests.
Changes in this CL fix TODOs about zero nonces and flaky tests. They
include the following.

ZERO NONCES

1. The KDF step now hashes public_key_part in addition to shared_key
to conform with the ECIES-KEM standard (see cipher.h for more
details).

2. The symmetric part of the hybrid cipher no longer samples a fresh
nonce. This saves both on randomness and communication. Instead, it
fixes an all-zero nonce which is also a secure way of implementing
AES-GCM *only* when using it for hybrid encryption.  (See
https://goto.google.com/aes-gcm-zero-nonce-security)

3. Updates SALT_LEN to 128 bits. As we're saving on nonce bandwidth,
I choose the conservative option of having a cryptographically strong
salt in the HKDF step.

This implementation is more secure than the previous one and saves on
at least 4 bytes of overhead for every PKE.

FLAKY TESTS
The previous implementation was failing on a small fraction of test
cases. This was caused by a padding error where 1-in-256 keys were
not 32 bytes long causing padding confusion. In an ideal deployment,
I would imagine, we discard keys that aren't 32 bytes long.

SHA256 --> SHA512
We also update the HKDF algorithm from using SHA256 to SHA512. This is
consistent with using AES256-GCM to provide 256-bit symmetric key
hardness.

Change-Id: I0c4643f860a831519a79dca4b2b2998af313cc0e