Pass by value if and only if you can later move instead of copy.
The reason the DataStore interface passes strings by value instead
of const ref is that it can eventually replace a copy with a move.
In this CL we change one parameter vector to const vector& because
we do not ever end up moving it and we move a parameter that is
passed by value that we are accidentally copying.

Change-Id: I1de376b524ce2ab1f173bcee09cbaa161b29e244