Rework the Analyzer to use the new Datstore interface.
In previous work we added a new Datstore interface to
replace the existing one, but we have not yet reworked
the analyzer to use the new interface. That is what
we do in this CL.

We delete the old interface and implementation and
we rework the Analyzer to use the new ones. The
key difference is that in the new scheme there is
an additional abstraction layer called ObservationStore
that sits on top of the Datastore. The ObservationStore
deals with Observations instead of the lower-layer
concepts of rows and tables. In the reworking the
Analyzer no longer deals directly with rows and tables
but rather interfaces with the ObservationStore and so
only works at the conceptual layer of Observations.

Practically speaking this means that a bunch of logic
was moved out of the Analyzer and into the ObservationStore.

Here is a table of the old and new stuff.

old                                   new
--------------                ---------------
store.h                       data_store.h
store.cc                      data_store.cc
mem_store.h                    memory_store.h
mem_store.cc                   memory_store.cc
bigtable_store_old.h          bigtable_store.h
bigtable_store_old.cc         bigtable_store.cc
analyzer_test.cc              analyzer_service_test.cc
schema.proto                  <deleted, no longer needed>
<no corresponding thing>      observation_store.h, cc

Change-Id: Iede15970c7f41bc7550466320512603dcfb3e0ee