Pass strings by const ref if they are not moved.
It is a good idea to pass a std::string parameter by value instead of
by const ref in cases where you can then avoid a copy inside of the
function. In this CL we fix a few cases in which we were passing a
string by value but in fact we were copying anyway inside of the function.
We change the parameters to be of type const std::string &.

Change-Id: I134802c81011e09d37625fdc96d01d140d20040f