Factories should return unique_ptr, not shared_ptr.
It makes more sense for a factory to return a unique_ptr because
the factory is making a new object and the receiver is the unique
owner of the new object. The receiver can always move the object
into a shared_ptr later if he wants.

Having a factory return a shared_ptr is misleading because it
seems to imply that multiple invocations whould return a shared_ptr
to the same underlying object, but that is not true.

Change-Id: Ie726f805f64af20d170b6feee546c5901f512baf