Some changes to the Datstore API to be more compatible with Bigtable.
After playing with the Datastore API against the real Cloud Bigtable
we have made a few changes to the API.

(1) We add a WriteRows() method that allows a client to send many rows
to Bigtable in a single RPC rather than repeatedly invoking
WriteRow().

(2) We implement DeleteAllRows() using a single RPC in the Bigtable
Admin API rather than an RPC per row to be deleted.

(3) We replace DeleteRows() which deletes an arbitrary range with
DeleteRowsWithPrefix() which deletes all rows with keys that have
a given prefix. This latter may be implemented with a single RPC
to Bigtable's Admin API rather than an RPC per row to be deleted.

Change-Id: I841678273391bcc8906b4ad802fc956e59b49f32