Adds support for running tests against the real Google Cloud Bigtable
We have a set of templated tests that use the abstract interface
DataStore. Currently we have two instantiations of these templates:
One for a MemoryStore and one for a BigtableStore that connects
to the Bigtable Emulator running locally on the test machine.

In this CL we add the necessary infrastructure for a third instantiation
of the templates--one the uses a BigtableStore that connects to the
real Google Cloud Bigtable.

- bigtable_cloud_helper.h contains BigtableStoreCloudFactory. This factory
is the template parameter that causes the tests to use the real
Bigtable. The factory invokes BigtableAdmin::CreateFromFlagsOrDie()
for the admin interface and BigtableStore()::CreateFromFlagsOrDie()
for the DataStore interface. Thus at test time we must pass in flags
that point to the cloud project and instance.

-bigtable_store_cloud_test.cc instantiates the template DataStoreTest using
the factory BigtableStoreCloudFactory.

- observation_store_cloud_test.cc instantiates the template
ObservationStoreAbstractTest using the factory BigtableStoreCloudFactory.

- cobaltb.py and test_runner.py have been enhanced to be able to pass on
the command line flags that point to the cloud project and instnace

- README.md has a new section that gives details about how to use this.

Change-Id: Ie8a38cd1f995481b55a616b068dcf01955ca2ce8