Adds AddObservationBatch() method to ObservationStore.
Recently we added a bulk add method to the underyling
DataStore API. In this CL we take advantage of this
by implenting a bulk add method at the ObservationStore
layer.

In our Cloud Bigtable test this means that when we
add a batch of 100 observations we need to make only
a single RPC to Cloud Bigtable instead of 100 RPCs.
This speeds up the test by a lot.

Change-Id: Ieda36a0ee4064ac364df4237bddd8ccf34a7da5a