Adds report.proto
We add protocol messages for storing reports.
A report is the final output of the Cobalt Analyzer.

These  messages are stored in Bigtable and we will
offer an API for querying the pages of a report,
for querying the ReportMetadata of a report, and
for querying the set of reports for a given
report config.

This CL contains only the proto messages not that API.

Change-Id: I6d3e3c39d438ba8677645ea34b8b59ff4f8bb5c4