Adds two new tables to the Datastore model.
Our abstract Datastore interface has the notion of tables.
Prior to this CL there were two tables:
kObservations
KReports.

In this CL we change the set of tables to:
kObservations
kReportMetadata
kReportRows

That is we split kReports into two tables. This is in preparation for
an upcoming CL in which we introduce the ReportStore to manage reports.

Change-Id: I05752aa6aaf60b01f16632f11baad1272bd62a04