Adds EncodingMixer
An EncodingMixer is responsible for coordinating the analysis of a set of
observations that are possibly heterogeneous with respect to their encodings.
The observations are aggregated into homogeneous groups, the appropriate
decoder/analyzer is applied to each group, and the analysis results are
combined into a final result.

In V0.1 of Cobalt we don't support heterogeneous reports. So currently the
role of the EncodingMixer is to aggregate the observations into groups in
order to make sure that there is only one group--i.e. that only one
encoding_config_id is being used--and then invoke the correct decoder/analyzer.

Change-Id: Ie10542a5d0c079d3758322a121a24341b75176ca