Adds ReportGenerator
ReportGenerator is responsible for generating individual reports. It is not
responsible for knowing anything about report schedules. It is not
responsible for figuring out which interval of days a report should analyze.
Those things are the responsibility of the ReportMaster.

In Cobalt V0.1 ReportGenerator is a singleton, single-threaded object
owned by the ReportMaster. In later versions of Cobalt, ReportGenerator
will be a separate service.

The relationship between ReportGenerator and EncodingMixer (from the
previous CL) is that ReportGenerator is responsible for
(a) Reading Observations from the ObservationStore
(b) Inovking EncodingMixer
(c) Writing reports to the ReportStore
Recall that EncodingMixer doesn't interact at all with the data stores.

ReportGenerator will in the future also have the responsibility for
understanding how to execute joint reports of two variables. But that
functionality has not been implemented yet.

Change-Id: I85168b5080da2f3a2957f0be7637e227977d5dfa