Introduces ReportMaster Service proto.
report_master.proto contains the definition of the public-facing
ReportMaster service.

ReportMaster service is used by Cobalt customers to generate one-off
reports, to query the set of reports that exist, and to fetch individual
reports.

Note that in the primary use of Cobalt report generation occurs
automatically on a regular schedule and there is no reason to request the
generation of a one-off report. Also reports are pushed to the user via
some TBD mechanism so there is in genral no reason for a customer to
query for reports.

Nevertheless this service is available for customers that wish to generate
one-off reports or query reports. This interface may be used to query and
fetch both one-off, manually-requested reports and regularly scheduled
reports.

Change-Id: I431d8463fce6fc0deb7011e6946c5a6502fd3af7