Adds report_internal.proto.
We are replacing the old report.proto which was intended only
for internal use with the union of
report_master.proto and report_internal.proto.

report_master.proto contains the public API.

report_internal.proto contains the additional bits that are
needed by the implementation but not exposed in the
public API. Notably it contains ReportId and VariableSlice
from the old report.proto which we still want to use in the
private implementation but don't want to expose in the
public API.

In a previous CL we added report_master.proto but we did not hook
it into the build. In this CL we complete the process by deleting
the old report.proto, introducing report_internal.proto and
fixing up the code to use the two new protos instead of the
old report.proto.

The design of the public API in report_master.proto caused the design of
the private implementation APIs in report_store.h and report_generator.h
to change a little:

- Because report_ids in the public API are opaque strings and we don't
expose the ReportId message in the public API, it was necessary to add
more information to ReportMetadata for the public API. But we still want
the older, smaller version of ReportMetadata for the private implementation.
So we add a new message called ReportMetadataLite for use in the private
implementation. This is what gets stored in BigTable.

- While designing the public API we renamed ReportMetadata.status to
ReportMetadata.state.

- While designing the public API we moved the ReportState enum out of
the ReportMetadata message and made it a top-level enum.

- When designing the public API it  was necessary to introduce the
notion of report "creation time"
as something distinct from report "start time", and at the same time
introduce the WAITING_TO_START state. When in the public API a user
invokes StartReport and specifies a ReportConfig with mutliple variables,
we create and start the first variable slice of the report but we
also create and don't start the other variable slices. This feature
of the publci API caused us to change the API of ReportStore a little.
We replaced the StartAssociatedSlice() method with two new methods:
CreateSecondarySlice() and StartSecondarySlice().

In the public API we use google.protobuf.TimeStamp to represent
timestamps because it is a Google Standard. In the private
implementation we still want to represent timestamps as numbers so that
for example we can make them part of a BigTable row key. But in order
to better accomodate the use of google.protobuf.TimeStamp we changed
the internal representation of timestamps from
uint64_t timestamp_millis to int64_t timestamp_seconds. That is we now
use seconds instead of milliseconds and signed ints instead of unsiqned ints.

Change-Id: Ife283b5c2242509a6553e841d28d2389b7beb6e6