Introduces AnalyzerConfig
AnalyzerConfig provides a convenient interface to the Cobalt configuration
system that is intended to be used by the Analyzer server processes.

This means we won't have to pass around three separate config objects
in the Analyzer anymore and also this centralizes the logic for
finding the config files using flags.

Change-Id: Iccdfd1a33a251927271c810507be89643daabfce