Adds implmentation of ReportMaster Service
This is an implementation of the gRPC API that we defined
in report_master.proto. Previously ReportMaster contained just
stub implementations and this CL contains the real implementation.

Mostly this just stiches together the functionality defined in
ReportStore and in ReportExecutor. Because the representation
of report IDs and ReportMetadata is different in the public
API than it is in the internal API we have to do some translation
between the two.

Change-Id: Ifb4114dd8e10b8048fbc09e949b838c920081212