Make report_master_service_cloud_test run faster.
In ReportMasterServiceAbstractTest::QueryReportsTest we add 210 reports
to the ReportStore in order to be able to test the QueryReports() method.

But the gRPC API does not offer an ability to start many reports at once
(because there is no reason to do that in general) and so in order
to add 210 reports we needed to call StartReport() 210 times. This is
fine for all configurations of the test except for the cloud configration.
With the cloud configuration we needed to make 210 RPCs from the local
computer to cloud Bigtable and this was making the test take
too long.

In this CL we fix the problem by adding some infrastructure that allows
us to write many rows to the ReportStore with a single RPC.

We introduce a class called ReportStoreTestUtils which may have more
uses in the future but as of now has one method: WriteBulkMetadata().

Change-Id: Ibd78e0704c1ede27ba39b24838bad7caaa677e75