Adds encoder/ShufflerClient
This is jsut a thin wrapper around the gRPC client to the Shuffler
allowing an Encoder client to optionally not deal with the details
of gRPC.

We also change the name of the CMake target for the Shuffler's
gRPC c++ bindings to be parallel to the name for the Analyzer's.

There are also two trivial unrelated changes in the encoder directory
for problems I found while testing:
- fixed a log message
- Make the envelope() accessor public so it can be used by the
test client when skipping the shuffler and sending directly to the
Analyzer.

Change-Id: Ie49a43bf9dd6f65d2c1cdd73016fb4c855699ec2