Adds the Cobalt test application.
This is the test client application that I used in the demo.
It can be used in three modes:
(a) In interactive mode. This is what I showed in the demo.
(b) In send-once mode. This will replace cgen.
(c) In automatic mode. This runs forever generating random Observations
and sending many RPCs. This is not yet implemented.

I do not replace cgen in the current CL. I wil do that in a follow-up.

Change-Id: Ic7b1eefe82aa2a5b9a503507c2b9c9eb472c8da0