Rename the EncryptionShcheme enum values.
(1) Name the real encryption scheme that Ananth invented
HYBRID_ECDH_V1 instead of PK_SCHEME_1.

(2) Replace UNKNOWN_SCHEME with NONE. Currently we are using
the EncryptedMessage message to hold plaintext bytes of
various serialized protocol buffer messages. Even after we
finish hooking up public key encryption in Cobalt it will continue to
be useful to *not* use encryption in some unit tests--namely unit tests
in which the encryption is completely orthogonal and for which dealing
with encryption would be a distraction from the thing being tested.
For that reason it is useful to allow an EncryptedMessage to contain
unencrypted bytes by explicitly designating a scheme called NONE that
does that. This allows us to be very explicit about when an EncryptedMessage
contains plaintext and will allow us to check for it accidentally being
used.

Change-Id: I366365da5dbb12384c3cc8f6bb8a85ba654a0bf6