Make |pem_root_certs| an optional argument to ShufflerClient.
The ShufflerClient constructor takes an argument called
|pem_root_certs| which is an optional string containing
a PEM encoding of root certificates.

Since this argument is optional and usually not needed we change
it to a C++ optional argument by changing its type to
be a const char* with a default value of nullptr rather than
it being a std::string.

Change-Id: I3895d2b112d77543f77da56069a6bec2a1f64d4e