Consoldiates method of starting the Cobalt processes on a dev machine.
We add the file tools/process_starter.py and consolidate the logic for starting
all of the processes into this file. Previously there were three places
that contained logic for starting a Cobalt process:
(a) cobaltb.py contained the logic for starting a process for the purpose of
manually starting them for the sake of, say, a demo.
(b) test_runner.py contained the logic for starting a process as part of
a unit test.
(c) Our end-to-end tests contain logic for starting a process as part of the
end-to-end tests.

Now there will be a single place used for all three purposes. I don't update
the end-to-end tests in this CL--that will come in a later cl.

(This CL also includes an unrelated fix to the CMakeLists.txt files to
resolve the fact that the builds on the continuous integration machine
are flaky because they depend on the Shuffler being built before the
report_master_client. I just noticed that while preparing this CL so
I through in a fix.)

Change-Id: I24e8cdde6aa9351451e4c2c46433dc735a7d78dd