Some fixes to Shuffler I found while playing with the demo.
I exercised the new Shuffler code a lot by playing around with the
demo pretty extensively. This CL contains the following miscellaneous
fixes.

- Fixed a bunch of log messages and VLOG levels. It turns out I use
VLOG level 3 for the demo so I wanted to make sure that the set of log
messages printed at that level were well-phrased and that they had the
right granularity. I reworded some messages and other messages that were
more useful for debugging I raised to V(4).

- There were some log messages that were just broken. For example there
were some cases where Info() was used but Infof() needed to be used.

- In the demo (or whenever FrequencyInHours is set to zero) there is no
need to disconnect and reconnect because we would be doing that every
3 seconds.

- The computation "if waittime == 0 { waitTime = minwaitTimeInS}" was not
working because waitTime is a float and comparison to 0 was failing.
In fixing this I decided it was better to just use the Duration type
to represent wait time. That is what it is for.

- I got rid of hard-coding "/tmp" as the location for the LevelDB directory.
There is no reason to hard-code this inside of the Go code. For the
purpose of testing and demo we can hard-code it inside the scripts that
invoke the Go code.

Change-Id: I0e339a48fbf79607d85111199ab4bf5a502ff235