Adds a non-interactive mode to the query_observations tool.
Previously query_observations was an interactive command loop that
allowed a user to query the contents of the ObservationStore.

Here we add the ability to send the flag -nointeractive and instead
the tool will perform a single query and print out the count of
Observations.

I plan to use this in the end-to-end test.

Change-Id: I49c23ac3e024f85fb2b62fa65d3d39413519bf6d