Renames config/registered to config/demo
The Cobalt configuration files in that directory are currently being used
for our manual demo and for our end-to-end test. Both locations
assume a particular contents of those files and must be kept
in sync with them. For that reason we rename the directory
appropriately and also we add a README.md file explaining
the situation.

Change-Id: Ia36c4fd3ed67afa9dd7c349785c61e00532710b1