Adds NoOp Encoding.
That is a way of sending raw unencoded values through the Shuffler
to the Analyzer.

This CL is part 1 of the work indicated in
https://fuchsia-review.googlesource.com/#/c/24266/.

In this CL we actually introduce the encoding and implement it on the
encoder. This includes no changes at all on the Analyzer so that
no reports can be generated that consume the NoOp encoding.

But this should be useful to Ilya's work because he will be able
to directly query the unencoded Observations from the ObservationStore.

Change-Id: I6f26d20b0ff2c7c20dc9a6f9fc54d411bed73b6c