Miscellaneous fixes that support running Cobalt on GKE.
- We stop explicitly linking against gflags. I don't yet
fully understand what is going on here but we seem to be
pulling gflags in indirectly through something else
(I think glog) and when we explicitly link against
gflags it causes problems on gke. Still investigating this
but or now the solution of not explicitly linking against
gflags seems to work.

- Stop linking against MemStore in non-tests. (And so start explicitly
linking against MemStore in tests.)

- As part of the above delete DataStore::CreateFromFlagsOrDie()
because we will never use the flag -for_testing_only_use_memstore.
We now only use MemStore in single-process unit tests. Therefore
in production code directly invoke BigtableStore::CreateFromFlagsOrDie()

- Improve some logging

- Get rid of default AnalyzerURI in Shuffler.

Change-Id: I1f07c3aa1730d51b9b1dfe2ad42f60b4e4e6b05e