Allow observation_querier to count arbitrarily many observations.
Previously observation_querier, when used on command-line mode,
was capable of making only a single query and thereby counting up
to a maximum of 10,000 observations. Now it queries iteratively and
so it is able to count arbitrarily many observations.

We also changed the name of QueryOnce() to CountObservations() to better
reflect the fact that its purpose is to count and not return the
actual observations.

Change-Id: I8fdf32bea1ad0fd574aaf80b1d9e070709e6c9c3