e2e test: Respect skipShuffler argument.
The e2e test had a bug where the skipShuffler argument was being
ignored. We don't actually use this argument so it wasn't a problem.
But it is useful to be able to have it working in case we want
to play around with it as was the case with Ananth's work.

Change-Id: I3f2c6d43c46b19f4723f9aec4cf57b2d5c4fe288