Shuffler: Fix sleep times.
The expression time.Sleep(time.Duration(dispatchDelayInS)) causes
a sleep of |disaptchDelayInS| *nanoseconds* instead of dispatchDelayInS
seconds. The correct expression is
time.Sleep(dispatchDelayInS * time.Second).

We discoverd this while helping Ananth try to increase throughput as part
of his experiment.

Our solution here is to use the time.Duration datatype whenever possible to
represent durations so as to avoid confusion.

We also increase batchsize from 100 to 1000 which is more appropriate.

Change-Id: I672e90576754c8ad21d60ca1ba9c0bf0f5b2e0d9