Adds the ability to run the e2e test against Cloud Bigtable.
We now support the following command syntax:
./cobaltb.py test --tests=e2e --bigtable_project_name=<project_name> --bigtable_instance_name=<instance_name>

This will cause all the Cobalt processes to start locally but the Bigtable Emulator will not start and instead
the local processes will all use the specified version of Cloud Bigtable.

This passes for me on my laptop using a Cloud Bigtable instance I have provisioned.
Before anyone else will successfully be able to use it I will have to provide instructions.
This will not be part of the automated tests.

Change-Id: Ie97eba6313cc552811e0ec0e41e22d7f9c8614ec