Adds tools for deploying Cobalt to Google Container Engine.
(1) Adds the following commands to cobaltb.py

deploy authenticate: Authenticate the local machine with GKE and
                     associate with a particluar GKE cluster
deploy show : shows the state of the cluster
deploy build: Builds new Docker images
deploy push: Pushes Docker images to the repository
deploy start: Starts one of the jobs
deploy stop: Stops one of the jobs
deploy upload_secret_key: Copies an Analyzer private key to a secure key
                          storage in the cluster
deploy delete_secret_key: Deletes a previously uploaded private key

(2) Introduces the file personal_cluster.json. This is a file that each
developer must maintain if he wishes to work with Cobalt on
container engine. It specifies details about his personal cluster.

(3) Adds the ability for the observation querier to query Cloud Bigtable.
This is useful for debugging Cobalt in the Cloud.

My next CL will be a large update to the Cobalt documentation so that other
people besides me will know how to do all of this.

Change-Id: I835f6a6d668f940700e334395f0526e09cb24023