Add AEAD decryption support for Shuffler.
This cl uses the underlying golang crypto/aes and crypto/cipher
packages to perform AEAD encryption and decryption for sealed envelopes
using a non-standard nonce of size 16 bytes. This code performs AES-128
bit encryption.

Noopcrypter is removed and replaced with a SymmetricCipherCrypter dummy
impl in the receiver code. In the next cl, actual private key is
extracted from the fingerprint provided in the request and used for
decrypting the envelope.

Change-Id: Icc23192097523f96a3ddaf27953c1c3f73bd2702