Adds functionality to bigtable_tool to delete data from stores.
We add two new commands to bigtable_tool
delete_observations and delete_reports.
These will delete all data from the Observation Store or
the Report Store respectively.

We expose these as top-level command via
./cobaltb.py bigtable delete_observations
./cobatlb.py bigtable delete_reports

We also invoke this from the end-to-end test in the case that
the end-to-end test is being run against Cloud Bigtable
(whether the Cobalt processes are running locally or in GKE).

In this way the end-to-end tests always start with empty
Bigtable tables.

Change-Id: I6b7c9dc339c37c03116fb34ff9ff28b8b7a95ada