Implements Cobalt's ECDH-based hybrid encryption scheme in Go.
The scheme has already been implemented in C++ for use on the Encoder and
Analyzer
(see https://fuchsia.googlesource.com/cobalt/+/master/util/crypto_util/cipher.h)

here we implement the same algorithm in Go for use on the Shuffler.

The symmetric part of the hybrid encryption was already done in a
previous CL and is not part of this CL. But notice that wherease in
C++ the symmetric encryption uses AES 256, here we use AES 128. This is
because we intend to go back and change the C++ implementation to AES 128.

The new ECDH code in this CL is all straightforward except for one thing:
The Go libraries do not implement the compressed encoding so we found
we had to implement that ourselves. We found an old CL that was
abandoned and never included in the standard Go library so we
cribbed off of that.
https://go-review.googlesource.com/c/1883/
Our code is in elliptic_util.go.

The code there looks good but agl@ decided to not allow it in the
standard library. In a private communication he wrote
"Compressed points with NIST curves are sufficiently obscure that I
don't expect the Go stdlib to support them, I'm afraid. (New things are
 generally using X25519/Ed25519 where "compressed"
points are standard.)

Part of implementing the "decompression" involves computing a square
root in the field Z/p where p=p256. The Go standard library includes a
generic function for computing square roots mod any prime using the
standard Tonelli-Shanks algorithm. In the case of p=p256
we know that p mod 4 = 3 and that allows us to use a simpler and
possibly more efficient algorithm to compute square roots. Our code for
that is in p256.go. I don't know whether our implementation is actually
more efficient so I have left a TODO to profile it. If not we will
just fall back to the generic implementation.

We also threw one unrelated piece of house-keeping into this CL:
We deleted the interface Crypter from crypto_util.go. There is no point
in that inteface since the SymmetricCipher and the HybridCipher implement
sligthly different interfaces.

Change-Id: I09cd23fc78c44817fbf090dd4eb3d03e3eb115fb