Miscellaneous cleanups.
This CL contains a few trivial clean-ups that I came across while
working on the String RAPPOR encoder. I am pulling them out
as a separate CL so as to make the RAPPOR encoder CL easier to review.

- Runs clang format on some files
- Fixes some lint warnings.
- Removes an extraneous "testing" namespace
- Adds a version of ClientSecret::GenerateNewSecret() that
takes an instance of crypto::Random. This allows an instance
of DeterministicRandom to be passed so that ClientSecretes
can be generated deterministically in tests.

Change-Id: Ied0f32b8a507bce18d91f00ce247b8e4a2edfc31