Adds Encrypted Message utility to Shuffler.
encrypted_message_util.go essentially contains a translation into Go
of the code in util/encrypted_message_util.cc. This is a layer on top
of the crypto library that knows how to work with our EncryptedMessage
proto. (see encrypted_message.proto in the root directory.)

The C++ version of EncryptedMessgeMaker is used in the Encoder client.
The C++ version of MessageDecrypter is used in the Analyzer.
The Go version of EncryptedMessageMaker is used only for tests.
The Go version of MessageDecrypter is used in the Shuffler.

Change-Id: I9e9e34035c573120ef3b809be4bd07d064a830a0