Fix computation of ECDH shared key in Go.
I didn't fully understand what the shared key in ECDH was. If
g^alpha is the public key and beta is the private key then
g^{alpha*beta} is the shared key.

Except that object is an element of an elliptic curve group and
the shared key is not an abstract mathematical object it
is a byte sequence. So which byte sequence is it?
It turns out it is the big-endian bytes, padded with leading
zeroes to length 32, of the x coordinate of that point.

This CL fixes that. We add two new functions computeSharedKey() and
padFieldElement(). We also move some logic from crypto_util.go to
elliptic_util.go. We also change the name of a constant in the C++
implementation. GROUP_ELEMENT_SIZE was a misleading name. Since the
shared key is a representation of only the x-coordinate a better name
might be FIELD_ELEMENT_SIZE. Instead I used ECDH_SHARED_KEY_SIZE.

Change-Id: Iffff8f6ad4c0422c95325fd8f6f800d9f487433b