Correctly decode PEM files in Go.
I had a misunderstanding of what a PEM encoding of a public or private
key was. This bug was not detected until I started working on the
interaction between the C++ code and the Go code. Previously
I had a unit test in which I was incorrectly creating a PEM encoding
and then incorrectly parsing the PEM encoding that I created.

Now I have a unit test in which I use a literal PEM string that was
created by the C++ code and so I can test that the new implementation
of PEM decoding in GO is correct. I don't implement PEM encoding in GO as
we have no need for it.

Change-Id: I21d60d55049d49b639381840285b6011a120ded3