Improve the protobuf build.
This CL makes no changes to any behavior but improves the way protocol
buffers are compiled in Cobalt. The motivation for doing this is that in
a different CL I wanted to add a reference from a proto message in the
report_master directory to a proto message in the config directory and it
was impossible to make this compile in both C++ and Go the way our
protobufs were being compiled.

- All protobufs are now compiled to C++ using the root directory as the
only element of the .proto search path. This means that all of the
generated code thinks of itself as being located at a  path which is the
relative location of the .proto file. This provides a consistent and
uniform way for all the generated C++ code to reference all the other
generated C++ code. All generated code is referenced using its full path
and all .proto files import other .proto files using their full path.

- In order to make the above happen it was necessary to stop using the
convenient CMake macro protobuf_generate_cpp and instead directly
embed the full invocation of the protobuf compiler into the CMake file.
The macro was not expressive enough for the change we wanted to make.

- The generated Go code for the .proto files located in the config
directory is now in the "config" package. This makes it consistent with
the way the other generated go code was.

- In order to make the above happen it was necessary to change the name
of the Shuffler's config package from "config" to "shuffler_config".

Change-Id: I4e23cd29fd07e8e373951efa9f4649a2ef4f10a4