Replace notion of "variable slice" with notion of "report type"
Previously Cobalt was hard-coded to be able to generate two types of
reports: a single-variable historgram report using RAPPOR or Forculus,
and a two-variable joint-analysis report. This restriction was embedded
into the protocol buffers and the APIs. A ReportID used to contain a
final field called VariableSlice which had the value 0, 1 or 2. This
field served two purposes: It gave the sequence number of the report
in its dependency group and it also described the type of the report:
0: First report to be generated. Analyzes the first variable.
1: Second report to be  generated. Analyzed the second variable (if there
is one)
2: Third report to be generated. Performs the joint analysis of two
variables (if there is a second variable.)

In this CL we break this tight association with these two types of report
by introducing the notion of a  "report type" and changing the format of
a report ID. A ReportConfig has a new field of type ReportType. So far
we have not added any new types of reports--there are still just
HISTOGRAM and JOINT--but now this is explicitly specified by an enum.

Also a ReportMetadata now specifies a ReportType and also a ReportMetadata
now specifies the list of variables indices which it analyzes.

Also we changed the internal APIs of the ReportMaster and ReportStore so
that instead of starting a "dependent slice" of a report they now talk about
starting a "dependent report" and instead of relying on the ReportSlice
field of a ReportId to determine which variables to operate on they now
are explicitly given a list of variables to operate on.

OLD SCHEME                               NEW SCHEME
---------                                -----------
report_id had a VariableSlice field.     report_id has a sequence_num field
Only 2 types of reports--implicit        extensible report types -- explicit
                                         ReportType is a field of ReportMetadata
					 ReportType is a filed of ReportConfig
					 varialbe_indices is a field of
					     ReportMetadata
ReportStore::CreateSecondarySlice()     ReportStore::CreateDependentReport()
ReportStore::StartSecondarySlice()      ReportStore::StartDependentReport()
                                        ReportStore methods take a
					|sequence_number| and a
					|variable_indices| argument to tell
					them the information that was
					previously implicit in the VariableSlice
					field of a ReportId.

Change-Id: If03badddf99999155246f3564321862e6c4add50