Introduce the notion of different types of report rows and analysis engines
This is a continuation of the work from
https://fuchsia-review.googlesource.com/#/c/31766/
in which we introduced the notion of ReportType.

As there are different types of reports there should be different types of
report rows. Prior to
https://fuchsia-review.googlesource.com/#/c/31766 there were two types of
reports, histogram reports and joint reports, and this was not extensible.
So previously there was a single type of ReportRow and it was designed to
be useful for both types of reports. A ReportRow contained two optional
values called |value| and |value2|. For the first marginal report |value|
was populuated, for the second marginal report |value2| was populated, and
for a joint report both |value| and |value2| were populuated.

Now we are moving away from a hard-coded set of report types to an
extensible set of report types. Consequently we need ReportRow to be more
flexible.

Here we change ReportRow to be a OneOf. Corresponding to our current set
of two ReportTypes we introduce two types to be included in the OneOf:
HistogramReportRow and JointReportRow. JointReportRow is so far a
place-holder as JOINT reports are not implemented. HistogramReportRow has
the fields previously in ReportRow, except that we no longer have a
need for |value2|. This is because we will never need to include 2 values
as a HistogramReportRow will not be used for a JOINT report, and we do not
need to encode the fact that a one-way marginal is for variable 2 in a
ReportRow as now ReportMetadata specifies which variable indices a report
is for.

This CL also contains a different but related change:
As there are now different types of reports we will need different types
of report generators.
We are already using the term |ReportGenerator| for a higher-layer that
does not depend on the report type so we introduce a new term
"analysis engine" to refer to the layer that will be different for
different types of reports. Currently there is only one
"analysis engine", namely the one that creates HISTOGRAM reports.

We rename |EncodingMixer| to |HistogramAnalysisEngine|. The EncodingMixer
contained the logic that was specific to HISTOGRAM reports but as we used
to have a hard-coded set of reports there was no need to indicate this in
the name of the class. Now that we have an extensible set of report types
we have re-branded EncodingMixer so as to indicate it is specifically
for the purpose of generating our HISTOGRAM reports.

Because ReportRow used to have a |value| and a |value2| it was necessary
to tell EncodingMixer which value to populate by passing it a |Variable|
object. As the new HistogramReportRow only has a single |value| and no
|value2|, we have removed the |Variable| from |HistogramAnalysisEngine|.

Change-Id: I64f5f4efcbda6c2255c9a96b9b7bf6121039b5e3